Move Parfor passes to parfor module and disable modified tests
