+++ numba/parfors/__init__.py
@@ -1,1 +1,2 @@
 from numba.parfors import parfor_lowering
+from numba.parfors import parfor_passes
